% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_lagged_cor.R
\name{evaluate_lagged_cor}
\alias{evaluate_lagged_cor}
\title{evaluate_lagged_cor}
\usage{
evaluate_lagged_cor(object, plot = TRUE)
}
\arguments{
\item{object}{a lagged_scatter_result class object.}

\item{plot}{return plot or not.}
}
\value{
A ggplot2 object.
}
\description{
evaluate_lagged_cor
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
