% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-fit_smoothing_parameters.R
\name{fit_smoothing_parameters}
\alias{fit_smoothing_parameters}
\title{Fit parameters for LOESS data smoothing}
\usage{
fit_smoothing_parameters(
  x,
  time,
  threads = 2,
  span_values = seq(0.1, 0.5, by = 0.01),
  degree_values = 1:2,
  progressbar = TRUE,
  seed = 123
)
}
\arguments{
\item{x}{A numeric vector of values to be smoothed.}

\item{time}{A numeric vector of timestamps corresponding to \code{x}. (POSIXct)}

\item{span_values}{A numeric vector of span values to test (default is seq(0.1, 0.5, by = 0.01)).}

\item{degree_values}{A numeric vector of degree values to test (default is 1:2).}

\item{seed}{An integer for the random seed (default is 123).}
}
\value{
A list of the best span and degree parameters for the LOESS data smoothing method.
}
\description{
This function fits the span and degree parameters for the LOESS data smoothing method
by testing different combinations of span and degree values and selecting the best
parameters based on the R² value.
}
\examples{
data("step_data", package = "laggedcor")
best_params <- fit_smoothing_parameters(x = step_data$step,
                                        time = step_data$time)
best_params

}
\author{
Zirui Qiang \email{zirui_qiang@outlook.com}

Minghang Li \email{minghli@ethz.ch}
}
