% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_cor_result.R
\docType{class}
\name{lagged_cor_result-class}
\alias{lagged_cor_result-class}
\title{An S4 class that stores the lagged correlation result}
\description{
An S4 class that stores the lagged correlation result
}
\section{Slots}{

\describe{
\item{\code{x}}{x}

\item{\code{time1}}{time1}

\item{\code{y}}{y}

\item{\code{time2}}{time2}

\item{\code{idx}}{idx}

\item{\code{all_cor}}{all_cor}

\item{\code{all_cor_p}}{all_cor_p}

\item{\code{shift_time}}{shift_time}

\item{\code{which_max_idx}}{which_max_idx}

\item{\code{which_global_idx}}{which_global_idx}

\item{\code{max_idx}}{max_idx}

\item{\code{max_cor}}{max_cor}

\item{\code{global_idx}}{global_idx}

\item{\code{global_cor}}{global_cor}

\item{\code{parameter}}{parameter}
}}

