% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-lagged_scatter_plot.R
\name{lagged_scatter_plot}
\alias{lagged_scatter_plot}
\title{lagged_scatter_plot}
\usage{
lagged_scatter_plot(
  object,
  x_name = "x",
  y_name = "y",
  which = c("global", "max"),
  hex = FALSE
)
}
\arguments{
\item{object}{a lagged_scatter_result class object.}

\item{x_name}{x_name}

\item{y_name}{y_name.}

\item{which}{global and max}

\item{hex}{TRUE or not}
}
\value{
A ggplot2 object.
}
\description{
lagged_scatter_plot
}
\examples{
data("object", package = "laggedcor")
lagged_scatter_plot(
  object = object,
  x_name = "Step",
  y_name = "HR",
  hex = TRUE, 
  which = "max"
)

lagged_scatter_plot(
  object = object,
  x_name = "Step",
  y_name = "HR",
  hex = TRUE, 
  which = "global"
)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
