% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laggedcor_logo.R.R
\name{laggedcor_logo}
\alias{laggedcor_logo}
\title{Show the logo laggedcor.}
\usage{
laggedcor_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\value{
A ASCII log of laggedcor
}
\description{
The laggedcor logo, using ASCII or Unicode characters
}
\examples{
laggedcor_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
