% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-utils.R
\name{laggedcor_packages}
\alias{laggedcor_packages}
\title{List all packages in the laggedcor}
\usage{
laggedcor_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include laggedcor in the list?}
}
\value{
laggedcor packages
}
\description{
List all packages in the laggedcor
}
\examples{
laggedcor_packages()
}
