% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lagged_correlation.R
\name{calculate_lagged_correlation}
\alias{calculate_lagged_correlation}
\title{calculate_lagged_correlation}
\usage{
calculate_lagged_correlation(
  x,
  y,
  time1,
  time2,
  time_tol = 1,
  step = 1/60,
  min_matched_sample = 10,
  progressbar = TRUE,
  all_idx = NULL,
  threads = 10,
  cor_method = c("spearman", "pearson")
)
}
\arguments{
\item{x}{Time-series data 1. A numeric vector.}

\item{y}{Time-series data 2. A numeric vector.}

\item{time1}{Time-series data 1. A time vector (POSIXct).}

\item{time2}{Time-series data 2. A time vector (POSIXct).}

\item{time_tol}{Time tolerance to match time-series 1 and time-series data 2. Unit is hour.}

\item{step}{Step to calculate lagged correlation.}

\item{min_matched_sample}{Minimum matched sample numner for one lagged correlation.}

\item{progressbar}{progressbar}

\item{all_idx}{all_idx}

\item{threads}{threads number}

\item{cor_method}{spearman or pearson.}
}
\value{
A lagged_cor class object.
}
\description{
calculate_lagged_correlation
}
\examples{
\dontrun{
data("heart_data", package = "laggedcor")
data("step_data", package = "laggedcor")

dim(heart_data)
dim(step_data)

x = step_data$step
time1 = step_data$time

y = heart_data$heart
time2 = heart_data$time

object =
  calculate_lagged_correlation(
    x = x,
    y = y,
    time1 = time1,
    time2 = time2,
    time_tol = 0.2,
    step = 1 / 60,
    min_matched_sample = 10,
    progressbar = TRUE,
    threads = 5,
    cor_method = "spearman"
  )
object
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
