% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_alignment_plot.R
\name{lagged_alignment_plot}
\alias{lagged_alignment_plot}
\title{lagged_alignment_plot}
\usage{
lagged_alignment_plot(
  object,
  x_color = "#631879FF",
  y_color = "#E377C2FF",
  x_name = "x",
  y_name = "y",
  which = c("global", "max"),
  x_limit = c(1, 1000),
  non_matched_point_size = 0.1,
  y_point_size = 1,
  x_point_size = 3,
  integrated = FALSE,
  add_connect_line = FALSE,
  add_point = FALSE,
  time_gap = 4
)
}
\arguments{
\item{object}{a lagged_scatter_result class object.}

\item{x_color}{x_color}

\item{y_color}{y_color}

\item{x_name}{x_name}

\item{y_name}{y_name}

\item{which}{global or maz}

\item{x_limit}{x_limit}

\item{non_matched_point_size}{non_matched_point_size}

\item{y_point_size}{y_point_size}

\item{x_point_size}{x_point_size}

\item{integrated}{integrated}

\item{add_connect_line}{add_connect_line}

\item{add_point}{add_point}

\item{time_gap}{time_gap in x axis.}
}
\value{
A ggplot2 object.
}
\description{
lagged_alignment_plot
}
\examples{
data("object", package = "laggedcor")

lagged_alignment_plot(object = object,
                      x_limit = c(1, 1000))
lagged_alignment_plot(object = object,
                      x_limit = c(1, 10000),
                      time_gap = 10)

lagged_alignment_plot(
  object = object,
  x_limit = c(1, 100),
  time_gap = 1,
  add_point = TRUE,
  x_point_size = 1,
  y_point_size = 1,
  integrated = TRUE,
  add_connect_line = TRUE
)

lagged_alignment_plot(
  object = object,
  x_limit = c(1, 100),
  time_gap = 1,
  add_point = TRUE,
  x_point_size = 1,
  y_point_size = 1,
  integrated = FALSE
)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
