% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-smooth_data.R
\name{smooth_data}
\alias{smooth_data}
\title{Perform Data Smoothing with Loess}
\usage{
smooth_data(
  x,
  time,
  span,
  degree,
  should_plot = FALSE,
  color_original = "blue",
  color_smoothed = "red"
)
}
\arguments{
\item{x}{A numeric vector of values to be smoothed.}

\item{time}{A POSIXct or POSIXlt object representing the time points of the x values.}

\item{span}{A numeric vector of span values to be used in the Loess smoothing.}

\item{should_plot}{A logical value indicating whether to plot the original and smoothed data.}

\item{color_smoothed}{The color to be used for the smoothed data in the plot. Default is "red".}

\item{color_orignal}{The color to be used for the original data in the plot. Default is "blue".}
}
\value{
A list containing the original and smoothed data, as well as the span values used.
}
\description{
This function performs data smoothing using the Loess method and opitonally plot
the original data and smoothed data to visualize the effect of different span
values on the smoothing process.
}
\examples{
data("step_data", package = "laggedcor")
x <- step_data$step
time <- step_data$time
span <- 0.1
degree <- 2

res <- smooth_data(x, time, span, degree, should_plot = TRUE)

}
\author{
Zirui Qiang \email{{zirui_qiang@outlook.com}}

Minghang Li \email{{minghli@ethz.ch}}
}
