% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plot.R
\name{time_plot}
\alias{time_plot}
\title{time_plot}
\usage{
time_plot(
  x,
  time,
  color = "blue",
  y_axis_name = "y",
  time_gap = 12,
  add_point = FALSE
)
}
\arguments{
\item{x}{x}

\item{time}{time}

\item{color}{color}

\item{y_axis_name}{y_axis_name}

\item{time_gap}{time_gap}

\item{add_point}{add_point}
}
\value{
A ggplot2 object.
}
\description{
time_plot
}
\examples{
data("step_data", package = "laggedcor")

x = step_data$step
time = step_data$time

time_plot(x, time)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
